Refactored client classes into separate modules
